requirements.txt
pyyaml
beautifulsoup
pytest-playwright
discord.py